requirements.txt
beautifulsoup4==4.11.1
httpx==0.23.0
aiohttp==3.8.1
pytest-aiohttp==1.0.4
